+++ Warcraft II/Warcraft II/j1Scene.h
@@ -62,7 +62,7 @@ public:
 	float up = false, down = false, left = false, right = false;
 	uint width = 0;
 	uint height = 0;
-	uint scale = 0;
+	float scale = 0;
 
 	// Player
 	bool god = false;
